updated feedback mechanism when testing mapping
